Moved managment code from init to a proper module.
Imported run_from_command_line, ManagementCommand and
ManagementTask in the init file.

Refs: #124
